requirements.txt
Flask==2.1.1
Flask_PyMongo==2.3.0
pymongo==4.1.0
python_bcrypt==0.3.2
